+++ test/game.py
@@ -4,7 +4,7 @@ from pyengine.Utils import Colors, Vec2
 from pyengine.Components import PositionComponent, SpriteComponent, PhysicsComponent, TextComponent, \
     ControlComponent, LifeComponent
 from pyengine.Entities import Tilemap, Entity
-from pyengine.Widgets import Checkbox, ProgressBar, Button, Entry, Label, Console, MultilineLabel
+from pyengine.Widgets import Checkbox, ProgressBar, Button, Entry, Label, Console, MultilineLabel, AnimatedImage
 
 
 class Game(Window):
@@ -23,6 +23,7 @@ class Game(Window):
         self.la = Label(Vec2(400, 450), "Contenu de \nl'entry", Colors.BLACK.value)
         self.console = Console(self, width=self.width)
         self.ml = MultilineLabel(Vec2(10, 450), "Chocolat\nVanille\nCARAMEL !", Colors.BLACK.value)
+        self.im = AnimatedImage(Vec2(50, 400), ["images/idle.png", "images/sprite0.png"])
 
         self.uisys.add_widget(self.check)
         self.uisys.add_widget(self.progress)
@@ -31,6 +32,7 @@ class Game(Window):
         self.uisys.add_widget(self.la)
         self.uisys.add_widget(self.console)
         self.uisys.add_widget(self.ml)
+        self.uisys.add_widget(self.im)
 
         self.esys = self.world.get_system(EntitySystem)