+++ Changelog.md
@@ -10,6 +10,7 @@
     - Console : Create console widget
     - MultilineLabel : Create Label widget for multiline
     - Label : Replace \n to void. Label doesn't support break lines
+    - AnimatedImage : Create Image widget which support multiple images
     - Unit Tests : Fix Vec2 and Color tests
     - Unit Tests : Add Console and MultilineLabel tests